feat(PITR): Fix main logical path of PITR restore

+++ api/issue.go
@@ -187,7 +187,7 @@ type PITRContext struct {
 	DatabaseID int `json:"databaseId"`
 	// After the PITR operations, the database will be recovered to the state at this time.
 	// Represented in UNIX timestamp in seconds.
-	PointInTimeTs int `json:"pointInTimeTs"`
+	PointInTimeTs int64 `json:"pointInTimeTs"`
 }
 
 // IssueFind is the API message for finding issues.
