+++ api/task.go
@@ -90,7 +90,7 @@ type TaskDatabasePITRRestorePayload struct {
 	ProjectID int `json:"projectId,omitempty"`
 	// After the PITR operations, the database will be recovered to the state at this time.
 	// Represented in UNIX timestamp in seconds.
-	PointInTimeTs int `json:"pointInTimeTs,omitempty"`
+	PointInTimeTs int64 `json:"pointInTimeTs,omitempty"`
 }
 
 // TaskDatabasePITRCutoverPayload is the task payload for PITR cutover.
